+++ stations/TWI/twiProtocol.c
@@ -65,11 +65,11 @@ static TWI_WORK             twiWork;
 static int readLineFromStation (WVIEWD_WORK *work, char *store, int msTime)
 {
     char        buffer[TWI_BYTE_LENGTH_MAX];
-    ULONGLONG   readTime = radTimeGetMSSinceEpoch() + (ULONGLONG)msTime;
+    uint64_t   readTime = radTimeGetMSSinceEpoch() + (uint64_t)msTime;
     int         retVal, timeToRead, done = FALSE, byteCount = 0;
 
     // we read until we see an "R" as a line terminator:
-    while ((timeToRead = (ULONG)(readTime - radTimeGetMSSinceEpoch())) > 0)
+    while ((timeToRead = (uint32_t)(readTime - radTimeGetMSSinceEpoch())) > 0)
     {
         retVal = (*work->medium.read)(&work->medium, &buffer[byteCount], 1, timeToRead);
         if (retVal != 1)
